refactor: reorg structure
